Move the weight_per_gas function into fp-evm
