+++ primitives/evm/Cargo.toml
@@ -17,6 +17,7 @@ serde = { workspace = true, optional = true }
 # Substrate
 frame-support = { workspace = true }
 sp-core = { workspace = true }
+sp-runtime = { workspace = true }
 sp-std = { workspace = true }
 
 [features]
@@ -29,5 +30,6 @@ std = [
 	# Substrate
 	"frame-support/std",
 	"sp-core/std",
+	"sp-runtime/std",
 	"sp-std/std",
 ]
